AK+Kernel: Added ObjectBuffer for aligned object storage
ObjectBuffer and the related ObjectArrayBuffer generic classes
provide the aligned storage for the (uninitialized) object of the
type specified in the template parameter. The classes provide
convenient access to the underlying aligned buffer, having the
alignment suitable for the object type, object pointer/constant
object pointer and reference/constant reference to the object
in-place constructed in the buffer.  No validation is performed to
check whether the object is constructed and can be (de)referenced.

The purpose of these components is to encapsulate the low-level
alignas logic into the dedicated helper components.